Updated domain module imports after MediaRepo, resolver moved to data module
