+++ core/domain/src/main/java/com/example/music/domain/model/GenreInfo.kt
@@ -3,7 +3,7 @@ package com.example.music.domain.model
 import android.util.Log
 import com.example.music.data.database.model.Genre
 import com.example.music.data.database.model.GenreWithExtraInfo
-import com.example.music.domain.util.Genre as GenreV2
+import com.example.music.data.mediaresolver.model.Genre as GenreV2
 
 private const val TAG = "GenreInfo"
