+++ core/domain/src/main/java/com/example/music/domain/model/SongInfo.kt
@@ -10,8 +10,8 @@ import com.example.music.domain.player.model.duration
 import com.example.music.domain.player.model.mediaUri
 import com.example.music.domain.player.model.title
 import com.example.music.domain.player.model.year
-import com.example.music.domain.util.Audio
-import com.example.music.domain.util.uri
+import com.example.music.data.mediaresolver.model.Audio
+import com.example.music.data.mediaresolver.model.uri
 import java.time.Duration
 import java.time.OffsetDateTime
