+++ core/domain/src/main/java/com/example/music/domain/usecases/FeaturedLibraryItemsV2.kt
@@ -3,7 +3,7 @@ package com.example.music.domain.usecases
 import android.util.Log
 import com.example.music.domain.model.FeaturedLibraryItemsFilterV2
 import com.example.music.domain.model.asExternalModel
-import com.example.music.domain.util.MediaRepo
+import com.example.music.data.mediaresolver.MediaRepo
 import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
@@ -17,16 +17,16 @@ class FeaturedLibraryItemsV2 @Inject constructor(
         Log.i(TAG, "Start fetching most recent albums and most recent songs")
 
         // albumItems should return albumRepo date created desc limit 5
-        val albumItems = resolver.mostRecentAlbums(5)
-        Log.i(TAG, "album items == $albumItems")
+        val albumIdsFlow = resolver.mostRecentAlbumsIds(5)
+        Log.i(TAG, "album items == $albumIdsFlow")
 
         // mediaItems should return songRepo date created desc limit 10
-        val mediaItems = resolver.mostRecentSongs(10)
-        Log.i(TAG, "media items == $mediaItems")
+        val mediaIdsFlow = resolver.mostRecentSongsIds(10)
+        Log.i(TAG, "media items == $mediaIdsFlow")
 
         return combine(
-            mediaItems,
-            albumItems
+            mediaIdsFlow,
+            albumIdsFlow
         ) { mediaIds, albumIds ->
             Log.i(TAG, "Building Featured Library from fetched IDs")
             FeaturedLibraryItemsFilterV2(
