+++ core/domain/src/main/java/com/example/music/domain/usecases/GetLibraryArtistsV2.kt
@@ -4,8 +4,8 @@ import android.provider.MediaStore
 import android.util.Log
 import com.example.music.domain.model.ArtistInfo
 import com.example.music.domain.model.asExternalModel
-import com.example.music.domain.util.Artist
-import com.example.music.domain.util.MediaRepo
+import com.example.music.data.mediaresolver.model.Artist
+import com.example.music.data.mediaresolver.MediaRepo
 import javax.inject.Inject
 
 private const val TAG = "Get Library Artist V2"
