+++ core/domain/src/main/java/com/example/music/domain/usecases/GetPlaylistDetailsV2.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import com.example.music.data.repository.PlaylistRepo
 import com.example.music.domain.model.PlaylistDetailsFilterResult
 import com.example.music.domain.model.asExternalModel
-import com.example.music.domain.util.MediaRepo
+import com.example.music.data.mediaresolver.MediaRepo
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
 import javax.inject.Inject
