+++ core/domain/src/main/java/com/example/music/domain/usecases/GetSongDataV2.kt
@@ -3,7 +3,7 @@ package com.example.music.domain.usecases
 import android.util.Log
 import com.example.music.domain.model.SongInfo
 import com.example.music.domain.model.asExternalModel
-import com.example.music.domain.util.MediaRepo
+import com.example.music.data.mediaresolver.MediaRepo
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
