+++ core/domain/src/main/java/com/example/music/domain/usecases/GetThumbnailUseCase.kt
@@ -1,12 +1,10 @@
 package com.example.music.domain.usecases
 
-import android.content.ContentResolver
 import android.graphics.Bitmap
-import android.util.Size
 import androidx.annotation.OptIn
 import androidx.media3.common.util.UnstableApi
-import com.example.music.domain.util.MediaRepo
-import com.example.music.domain.util.toAlbumArtUri
+import com.example.music.data.mediaresolver.MediaRepo
+import com.example.music.data.mediaresolver.MediaRepo.Companion.toAlbumArtUri
 import javax.inject.Inject
 
 class GetThumbnailUseCase @Inject constructor(
