+++ core/domain/src/main/java/com/example/music/domain/usecases/GetTotalCountsV2.kt
@@ -1,7 +1,7 @@
 package com.example.music.domain.usecases
 
 import android.util.Log
-import com.example.music.domain.util.MediaRepo
+import com.example.music.data.mediaresolver.MediaRepo
 import javax.inject.Inject
 
 private const val TAG = "Get Total Counts V2"
